refactor pipeline to match toast pipeline tools
